refactor: Update docking function name to dockm8_docking
