+++ scripts/docking/smina.py
@@ -12,7 +12,7 @@ from tqdm import tqdm
 dockm8_path = next((p / 'DockM8' for p in Path(__file__).resolve().parents if (p / 'DockM8').is_dir()), None)
 sys.path.append(str(dockm8_path))
 
-from scripts.utilities import (
+from scripts.utilities.utilities import (
     delete_files,
     printlog,
 )
